Add App.js
