+++ src/App.js
@@ -1,0 +1,13 @@
+import Calculator from './components/Calculator';
+import Title from './components/title';
+
+function App() {
+  return (
+    <main>
+      <Title />
+      <Calculator />
+    </main>
+  );
+}
+
+export default App;
