+++ src/index.js
@@ -1,27 +1,15 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import Calculator from './components/Calculator';
+import App from './App';
 import reportWebVitals from './reportWebVitals';
-
-const Title = () => {
-  useEffect(() => {
-    document.title = 'Calculator';
-  });
-  return (
-    <section hidden><div>Lets do some Math</div></section>
-  );
-};
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <Title />
-    <Calculator />
+    <App />
   </React.StrictMode>,
 );
-
-Title();
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
